+++ src/app/login/login.component.ts
@@ -56,27 +56,29 @@ export class LoginComponent implements OnInit {
             this.toastr.error(err.error.err, "Error", {
               timeOut: 5000
             });
-            // return of([]);
-            return throwError(err);
+            return of("error");
+            // return throwError(err);
           })
         )
         .subscribe(res => {
-          this.userService.loginUserSocket(res.user);
-          if (this.cookieService.check("currentUser")) {
-            this.cookieService.delete("currentUser", "/");
-          }
-          res.user.home.latitude = parseFloat(res.user.home.latitude);
-          res.user.home.longitude = parseFloat(res.user.home.longitude);
-          res.user.work.latitude = parseFloat(res.user.work.latitude);
-          res.user.work.longitude = parseFloat(res.user.work.longitude);
+          if (res != "error") {
+            this.userService.loginUserSocket(res.user);
+            if (this.cookieService.check("currentUser")) {
+              this.cookieService.delete("currentUser", "/");
+            }
+            res.user.home.latitude = parseFloat(res.user.home.latitude);
+            res.user.home.longitude = parseFloat(res.user.home.longitude);
+            res.user.work.latitude = parseFloat(res.user.work.latitude);
+            res.user.work.longitude = parseFloat(res.user.work.longitude);
 
-          this.cookieService.set(
-            "currentUser",
-            JSON.stringify(res.user),
-            null,
-            "/"
-          );
-          this.router.navigate(["/"]);
+            this.cookieService.set(
+              "currentUser",
+              JSON.stringify(res.user),
+              null,
+              "/"
+            );
+            this.router.navigate(["/"]);
+          }
         });
     }
   }
